Move Database class to Infrastructure
